Произведена модуляция, подготовка к выполнению второго дз
