+++ js/data.js
@@ -1,0 +1,10 @@
+const NAMES = [
+  'Вика', 'Антон', 'Даша', 'Анастасия', 'Виктор', 'Николай', 'Женя'
+];
+
+const COMMENTS = [
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.'
+];
+
+export {NAMES, COMMENTS};
